requirements1.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cors==1.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-apscheduler==0.6.2
django-cors-headers==4.2.0
django-filter==23.3
django-import-export==3.3.7
django-model-utils==4.3.1
django-notifications-hq==1.8.3
django-phonenumber-field==7.1.0
django-simple-history==3.3.0
django-twilio==0.14.3.2
django-vite-react==0.0.3
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.12.2
frozenlist==1.3.3
future==0.18.3
geographiclib==2.0
geopy==2.4.0
gevent==23.7.0
googlemaps==4.10.0
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
MarkupPy==1.14
MarkupSafe==2.1.3
mssql-django==1.3
multidict==6.0.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
phonenumbers==8.13.16
Pillow==10.0.0
pycparser==2.21
PyJWT==2.7.0
pyodbc==4.0.39
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tablib==3.5.0
tldextract==3.4.4
twilio==7.17.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
